Scripts/services/energy/requirements.txt
python-kasa==0.5.3
